Improve install/zsh.sh

+++ install/zsh.sh
@@ -28,12 +28,15 @@ source "${ROOT}/install/common.sh"
 # Export main environment variables for ZSH
 export ZDOTDIR="${XDG_CONFIG_HOME}/zsh"
 export ZSH="${XDG_DATA_HOME}/zsh/.oh-my-zsh"
+# Custom directory
+ZSH_CUSTOM="${ZSH}/custom"
 
 DOT_ZSH="${DOTFILES}/zsh" # Alias for dotfiles zsh
 
 # Install {{{
 
 install_zsh() {
+    # MacOS comes with zsh as default shell, however keep it.
     if ! command_exists zsh; then
         brew install zsh
         info "Zsh installed."
@@ -50,7 +53,7 @@ set_zdotdir() {
     echo "export ZDOTDIR=\"\$HOME/.config/zsh\"" |
         sudo tee -a /etc/zshenv
 
-    msg "Succesfully set \$ZDOTDIR"
+    info "Succesfully set \$ZDOTDIR"
 }
 # }}}
 
@@ -59,7 +62,7 @@ set_zdotdir() {
 # Create XDG_CONFIG_HOME link
 link_xdg() {
     force_remove "${DOT_ZSH}" "${XDG_CONFIG_HOME}/zsh"
-    msg "Zsh dotfiles linked to XDG_CONFIG_HOME"
+    info "Zsh dotfiles linked to \$XDG_CONFIG_HOME"
 }
 
 # Create personal soft links
@@ -69,7 +72,7 @@ link_personal() {
 
     if [ -f "${source_file}" ]; then
         force_remove "${source_file}" "${ZDOTDIR}/.private.zsh"
-        msg "Linked personal files"
+        info "Linked personal files"
     else
         warn "${source_file} not exists"
     fi
@@ -79,10 +82,9 @@ link_personal() {
 # oh-my-zsh {{{
 
 install_oh-my-zsh() {
+    msg "Installing .oh-my-zsh"
     # Fresh install: Remove if exists
-    if [[ -d "${ZSH}" ]]; then
-        rm -rf "${ZSH}"
-    fi
+    [ -d "${ZSH}" ] && rm -rf "${ZSH}"
 
     wget --no-check-certificate \
         https://raw.githubusercontent.com/ohmyzsh/ohmyzsh/master/tools/install.sh \
@@ -90,7 +92,14 @@ install_oh-my-zsh() {
 
     #   --unattended: sets both CHSH and RUNZSH to 'no'
     ZSH="${ZSH}" sh /tmp/install.sh --unattended
+
+    # somehow .oh-my-zsh overwrites my .zshrc
+    cp "${DOT_ZSH}/.zshrc.main" "${DOT_ZSH}/.zshrc"
+
+    info ".oh-my-zsh installed."
 }
+
+# }}}
 # }}}
 
 # Customs {{{
@@ -99,50 +108,45 @@ install_oh-my-zsh() {
 custom_plugins() {
     msg "Installing custom plugins"
 
-    ZSH_CUSTOM="${ZSH}/custom"
-
     # zsh-syntax-highlighting custom plugin
     # https://github.com/zsh-users/zsh-syntax-highlighting/blob/master/INSTALL.md
     git clone "https://github.com/zsh-users/zsh-syntax-highlighting.git" \
-        "${ZSH_CUSTOM:-~/.oh-my-zsh/custom}/plugins/zsh-syntax-highlighting"
+        "${ZSH_CUSTOM}/plugins/zsh-syntax-highlighting"
 
     # zsh-autosuggestions
     # https://github.com/zsh-users/zsh-autosuggestions/blob/master/INSTALL.md#oh-my-zsh
     git clone "https://github.com/zsh-users/zsh-autosuggestions" \
-        "${ZSH_CUSTOM:-~/.oh-my-zsh/custom}/plugins/zsh-autosuggestions"
+        "${ZSH_CUSTOM}/plugins/zsh-autosuggestions"
 
     # cd-gitroot
     # https://github.com/mollifier/cd-gitroot
     git clone "https://github.com/mollifier/cd-gitroot.git" \
-        "${ZSH_CUSTOM:-~/.oh-my-zsh/custom}/plugins/cd-gitroot"
+        "${ZSH_CUSTOM}/plugins/cd-gitroot"
 
     # zsh-completions
     # https://github.com/zsh-users/zsh-completions
     git clone "https://github.com/zsh-users/zsh-completions" \
-        "${ZSH_CUSTOM:=~/.oh-my-zsh/custom}/plugins/zsh-completions"
+        "${ZSH_CUSTOM}/plugins/zsh-completions"
 
     info "Installed custom plugins."
 }
 
-# Install custom themes
+# Link custom themes
 custom_themes() {
-    msg "Installing custom themes"
-    for theme in "simple" "gallois"; do
-        force_remove \
-            "${DOT_ZSH}/oh-my-zsh/custom/themes/${theme}-custom.zsh-theme" \
-            "${ZSH}/custom/themes"
-    done
+    mkdir -p "${ZSH_CUSTOM}/themes"
+    force_remove "${DOT_ZSH}/oh-my-zsh/custom/themes/"* "${ZSH_CUSTOM}/themes"
+	info "Custom themes completed."
 }
 
 # Link custom completions
 custom_completions() {
     mkdir -p "${ZSH}/completions"
     force_remove "${DOT_ZSH}/oh-my-zsh/completions/"* "${ZSH}/completions"
-    msg "Cusom completions completed."
+    info "Custom completions completed."
 }
 # }}}
 
-# TODO: Need to be tested
+
 main() {
     install_zsh
     set_zdotdir
